# Breaking change: Improve save() performance by skipping index creation