@ShaneHarvey 2022-10-27
The tests pass locally, @bagerard this is ready for you consideration. Could you approve the github workflow?